Check plugin dependencies on generation
The check_IdeaPlugin_Dependencies rule doesn't take into account
runtimes of used languages because it doesn't have that information in
the model.  So if a model uses a language that is not otherwise included
as a dependency, no errors were reported in the editor but this missing
dependency would lead to ClassNotFoundExceptions later on.

The information is available during the generation time after modules
are loaded, so perform a check at that time as a last resort.