MPS-23122 facilitate language re-use by aggregation
Let structure aspect tell what external languages it depends from and in what fasion.
Since this knowledge is specific to structure aspect, it's responsible to report it. Though at the moment 'extends' is exposed at module level, we likely to change this as we move forward.
Dependency on a module level means we need special treatment of 'Language' module and we sort of bind 'Language' module with 'structure' aspect.
Much like editor aspect shall expose its own dependencies without demand for its langauge to extend another.
'Extends' dependency between language modules has to become irrelevant for purposes when one needs to find out set of languages for use in an editor, and, perhaps, fade away for deployed modules.