ImmatureReferences: nested concurrent maps to thread-local storage
Given the fact we use stored references by value, it's odd to rely on their volatile attribute to group them. Reference's model is null when a free-floating node is created, and != null once node is attached to a model, hence remove() has had hard times matching the reference (well, it didn't match them at all, eventually). The only beneficiary of model-wise grouping of references is repository listener that reacts to model removal, which is, to be honest, quite rare operation in a regular model lifecycle.