StaticReference: do not attempt to mature reference on access
There's no apparent reason to hurry up 'maturing' a reference. The idea is to leave it up to external code (MA+IR) or to leave nodes as is as long as it's custom model and no MA+IR around.
The logic to make references 'mature' ion read had been introduced in 5ac3704 with no indication of thorough judgement.