Correcting the code in the ChangesTestBase (no asserts in before and after sections + removeing unnecessary SwingUtilities uses)
Also the model removal must be executed using the null requestor since
otherwise the mps vcs subsystem does not receive vfs notifications.
How that test worked earlier is a mystery to me.
The whole test code must be reviewed and rewritten.
Using an infinite number of #waits is a bad style.