Implemented common mechanism for migrating enum property attributes.
Migrated property attributes now distingushed by enumPropMigrated flag;
for newly created attrbiutes enumPropMigrated=true, for old attributes
enumPropMigrated=false. EnumUsagesMigration#migrateEnumPropertyAttribute
ensures that given attribute is for enum property and not yet migrated,
and if so then marks this attribute as migrated and gives reference
to a migrated property declaration.