build lang: fix for fetch external jar dependency scenario
This is to cover rare and questionable scenario, when there's module A with 'jar' dependency that uses external location _within the same project_, and module B that depends on A.
B.fetchDependencies() discovers A's jar, but its getDependencyTarget() == null as both A and its external jar live under the same root.
Though seems reasonable for B not to care about mutula location of A and its jar (rather about that of the jar and B itself), the scenario happens when A references external location from its own layoyt, which is likely not quite correct.
We use this sort of dependencies for bootstrap modules (where classes of a module go to a jar in lib/).