fix MPS-29666 Idea plugin: Migration starts on new project.
Now, we have additional ways to initialize configuration, so there's no obvious condition that will allow to detect exactly those situations we need. So, for now we just do a fix that looks at least safe. A correct fix includes restricting access to the module descriptor and changing the module itself (or, maybe through it's write-handle). In addition, there should be no such thing as "updating module descriptor"