Tricky weaving into Language.java is accomplished.
Now we are able to generate 'feedback' from three different places.
Though the generation plans were impossible to reuse in my scenario,
still it is remarkable that such a complex task could be performed in MPS.

Furthermore, I have discovered two better ways to generate into
Language.java.
One is a only little better than 'switch' -- you can weave directly into the #createAspect
method, without all this confusing passing of the model parameter via
switch.

The problem is that even with this solution we still need to have those superfluous imports in
the model '@descriptor'.

In order to get rid of imports, we'd better use the common mapping label in
the generator of 'lang.descriptor' so that every aspect simply marks its
generated descriptor class with that common mapping label.
The generator of 'lang.descriptor' could loop himself through the models
of the module and get all the info from the generator extensions.
No need to have the dependencies reversed between the aspect generators.