Introduce getSupertypes() method it Type's behaviour and provide impl for BL types.
In order to get a list of proper direct supertypes it's no longer necessary
to run a query against the type system. Typechecker is no longer maintaining
the list of nominative supertypes, the rule engine is responsible for resolving
any conversions, but there still might be a need to get the list of supertypes.
This method exists in response for this need and solves the problem for BL only.