DefaultEditorMessage to suggest methods with openapi.EditorComponent
EditorMessage API uses nodeEditor.* classes which is fine provided it's part of the same package. However, for subclasses from outside of the editor, it's not convenient to cast openapi.EC to implementation class every time they need to access some shared impl from DEM.
Besides, getCellForParentNodeInMainEditor is of no interest for editor runtime and is functionality specific to DEM implementation, no reason to expose it in EditorMessage interface, nor to use from outside.