MPS-30257 Editor is not closed for deleted root
Since NodeVirtualFileSystem is registered via 'virtualFileSystem' extension point instead on application component in VirtualFileManagerImpl constructor `myVirtualFileListenerMulticaster.getMulticaster()` is not added as VirtualFileListener to it.
This is reason why FileEditorManagerImpl.MyVirtualFileListener doesn't get notification that file RootNodeVirtualFile is deleted and doesn't close editor tab.

There is no way to get EventDispatcher<VirtualFileListener> myVirtualFileListenerMulticaster to add its multicaster as listeners, but now it is connected to MessageBus with VirtualFileManager.VFS_CHANGES topic. So the best way is to use it and send events through MessageBus.

For now implemented only for events used in FileEditorManagerImpl.MyVirtualFileListener.