MPS-30663 Path variables used in modules.xml instead of $PROJECT_HOME$
Nothing grand has happenned -- just using the project macro provider when saving
the project from IDE.
That provider first checks for the intellij system PROJECT_DIR macro,
thus the project paths which are in the project are guaranteed to be
persisted with PROJECT_DIR macro, not the user-defined one.

Loading still works with user-defined macros.