MPS-30736 transient configuration bean for short-term editing activities
Responsibilities of MPSConfigurationBean and SolutionDescriptor intersect to great extent. There's plugin code that deals with facet configuration through the Bean class, general MPS code that deals with SolutionDescriptor, and MPSFacet configuration page that uses both (reuse of ModelRootContentEntriesEditor required ModuleDescriptor). Now, module's SolutionDescriptor is the primary source of all non-plugin specific settings, and the bean class is an editing handle, which is not supposed to live long enough to face modification of the SD.