Project modules: react to descriptor changes much like SLibrary does.
Eventually, would like to get rid of hidden assumption that a file AM is associated with could be used to re-load the module. There's no guarantee reading the same file yields the same module (or any module at all, for the matter). Code that manages module reloading based on FS changes has to be external to the module.