JavaModuleFacet: explicit classes/ CP for stub modules
Of 93 solutions with compileInMPS==false, I've updated not only those with hand-written source/ root but also with MPS-generated sources in source_gen/, as we need CP of these modules to compile other MPS modules (both MPS and IDEA-compiled).

Some of these classes/ are mentioned in bootstrap idea.additional.classpath.txt, so it's of no real value, yet I'd like (a) to be explicit about their CP; (b) I'd like to introduce support in lang.build that uses this knowledge to better manage build layout (BM_ModuleDependency to reference JavaModule in addition to JavaLibrary)
Changes to modules like ide/ui, that are not part of bootstrap, are more relevant.