lang.build: mechanism to tell java library location from JavaModule dependency
When an MPS module depends from classes compiled in IDEA, we usually go with a jar dependency (and overriden location, if necessary).
However, as we don't prepare jar artifacts for all such cases, it's handy to specify location with classes directly in MD.
Therefore, build lang needs a mechanism to ensure these classes are properly referenced in dependencies.