build lang: no reason to avoid generic ModuleSources processing for a single-jar module
There could be ModuleResources that we'd like to get deployed, let alone it's odd not to invoke switch in this scenario.
Didn't add switch_ModuleSourcesToSourceLocation as I would like to have this call as part of 'sources of' template, rather than to place it over and over again in various places (it's no op for a module without java sources at the moment anyway)
Could not do the same for DevKit branch as it's still _AbstractModule, not _Module (pending change).