MPS-31295 respect languages that are part of build sequence
When a module uses a language which is part of the same build sequence, make sure dependencies graph reflect this to help cluster modules properly
This also alters MPS-30639 hotfix. If the language in use is the only one for a model, adding it to usedLanguage without extended didn't help to build a model (despite the proper module build order, no other make facets were injected without 'extends lang.core').