MPS-31417, MPS-31212 ModuleClassLoader is disposed and not operable
We change the logic of PluginLoaderRegistry to avoid several flaws we
have found there.
For instance, two-step EDT execution seems to be incorrect.
Also the queue of post-runnables is back at the state as it was a month ago.
The most correct way (without memory leaks) is to have per-module cl
tracking in PluginLoaderRegistry (to be implemented in 201)

We use special IntelliJ progress to avoid freezy behavior on the plugin reload.