MPSSPRT-226 Regression - Tooltips disappear too quickly - MPS 2019.2.4
Happened due to change in IdeTooltipManager:
89741740: git: IDEA-207700 Branches Popup: add tooltips for incoming/outgoing markers

In MPSToolTipManager#mouseMoved first if check was not working for some time, but did not affected functionality:
there is another check in MPSToolTipManager#showToolTip(String, JComponent, Point, boolean) to avoid recreation of the same tooltip.
But after platform change, tooltip is now closed on mouse move, because IdeTooltipManager#isTooltipDefined returns false: own interface TooltipComponent#getMPSTooltipText is used instead of JComponent#getToolTipText(MouseEvent) from JDK.

It was considered that MPSToolTipManager is outdated and everything was rewritten to use directly IdeTooltipManager.