lang.build: introduce a hack to dela with huge lists of MPS.IDEA dependency jars
For build language, it's compilation that is vital, while the logic what we include in MPS.IDEA solution deps is not so obvious.
Unless we settle this down, I don't see a reason to put extra effort to keep the list of jars inside MPS.IDEA.msd and in its mpsBootstrapCore counterpart synchronized.
Perhaps, with reasonable/slim list of jars in MPS.IDEA.msd, we can use regular lang.build approach and list all the jars.