MPSSPRT-211 ECJ fails to compile some code using xml parsing api from JDK
The initial problem: https://stackoverflow.com/questions/51094274/eclipse-cant-find-xml-related-classes-after-switching-build-path-to-jdk-10
Simply speaking ecj-4.10 has sporadic compilation errors ('Cannot resolve...') when there is a package in classpath which is visible from two different modules.
The newer ecj-4.13 (which has replaced in the distribution ecj-4.10 in 2020.1) yields the correct error message https://bugs.eclipse.org/bugs/show_bug.cgi?id=536928#c32
And the error is permanently reproducible, which is great.

In our case the problem package is org.w3c.dom.* and the problem modules
are 'java.xml' (which is bundled with JDK and ok) and the module 'unnamed'.
The package is distributed in lib/platform-util-ui.jar, which has 3 classes in org.w3c.dom package, that appeared
as a result of a few fixes of the similar problems (https://youtrack.jetbrains.com/issue/IDEA-176725, https://youtrack.jetbrains.com/issue/IDEA-193340).

Note that for instance xml-apis.jar would give the same problem, but
fortunately the xerces jars included in the platform bootstrap cp, but
not exposed as MPS stubs.
Though platform-util-ui.jar is exposed as a stub in the solution
MPS.IDEA, and that is a problem.
Here I just split this jar in two, hoping that IJ will do the same in 2020.
These 3 classes are just to fix the xerces hell in the IJ platform, and thus I do not see
any reason for MPS to expose them as stubs.

I do not see any rational solution for IJ to address batic--xerces
compatibility problems and not to have a conflicting set of jars in its 'lib' folder.
Meaning that the package org.w3c.dom is still contained in several
library jars and MPS strategy is to leave each particular scenario up to the user to solve
(providing correct stub solutions for the libraries which are used and
setting up an IJ plugin)