MPSSPRT-241 Editor ignores OS when suggesting shortcut ALT + Enter
Temporary presentation fix for 2019.3.
In 2020.* proper fix will be introduced: currently shortcut is registered in jetbrains.mps.nodeEditor.IntentionsSupport and it is instance of javax.swing.AbstractAction. For proper fix it need to be rewritten as com.intellij.openapi.actionSystem.AnAction so in LightBulbMenu shortcut can be found by actionId. This will also allow to change shortcut from Keymap settings what is not currently possible.