MPSSPRT-253 Facet in Editor Distribution Gets Disabled After "Invalidating Caches & Restarting"
The problem: the facet factory is not there when we read project modules.
Even though the facet factory is registered from library application plugin
(j.m.lang.plugin.standalone) and the module is from the project, there
are no guarantees that lib extensions are loaded when the project is
being opened.
The trend here is toward independence, so I expect these two processes
(loading of sources and loading of runtime objects) to diverge even
more.
That being said, if we support facet factory registration in
lang.plugin.standalone.ApplicationPlugin, we need to use some lazy or blocking
mechanism to be able to read the module facet using the facet factory.
Blocking is bad, I want mps-plugins lifecycle to be totally asynchronous
and (hopefully) outside of EDT. Thus lazy is our choice.

I see that instead of allowing custom MPS API invocations inside of lang.plugin
structures we could better create generateable language constructs (like
'module facet', 'model root', 'custom persistence') and handle all the
problems by ourselves (problems like MPSSPRT-253 and many more).
This way the access to MPS API extension points must be much more
restricted, as the majority of them have a lifecycle not compatible with
the lifecycle of mps plugins.
However I estimate this as a deeper task, so not here.

Done:
* Defining 'unknown_facet' in MPS -- now if a facet is not parsed, the user
   will have some idea about it, at least.

* Adding a callback for delayed facet instantiation -- not sure that API is to be this way, but it is precisely when I need now.