MPS-31638 NPE in FileBasedIndexImpl.indexUnsavedDocument
The method erroneously doesn't check project for null when processed index is deemed 'psi-based' (has to be fixed with a distinct patch for IDEA). However, there are not psi indexes in MPS. To tell FileBasedIndexImpl not to assume index to be 'psi-based' by default, we override it with a system property (see FileBasedIndexImpl.isPsiDependentIndex).