Avoid rasing an exception on "uknown" session release. MPSSPRT-269
Under some yet-to-be discovered conditions, the method `WorkbenchTypecheckingController.sessionReleased()`
is called with a session that has already been released, and
since an exception is thrown at this point, this makes the inspector unusable.
This change alters the behaviour to reporting an error instead, so that
the root cause of the problem can be investigated, yet keeps the UI responsive.