MPS-21683 It's possible to open find action window, but impossible to use it
Because there is no reference to java action concept in actions system, wrapper action is used with the same shortcut as for original action.
For some reason this leads to issue problem.

As quick fix before release - use this action only in MainMenu (reference was created for it) and allow original action to be used anywhere else.

Long term solution would be new Reference type of ActionGroup entry.