MPS-32637 flushing properly the reloads, not scheduling a new reload
double invocations of EditableSModelBase#save led to doubling
disk-memory conflict dialog

the method #flushAllPendingReloads is added to ReloadManager

Not sure what is meant by the original #flush method implementation.
Also no understanding whatsoever why the logic is so complex (separate
queue, some tasks) when eventually everything goes in awt thread.
Probably it is worth to drop these complexities but not this time.