MPSSPRT-286 Adding a check for headless evnironment where it is possible
Ever since the change in EnvironmentBase, where we stopped setting test mode
every time we are headless the checks ```Runtime.isTestMode()``` written
all over ui platform extensions in MPS, stopped working.
The solution is to extend these conditions with a headless check.

Artem says it makes sense to facade the headless mode in MPS, but I do
not see the need for it in bugfix.
This is something we will do one day, though, when we will need this.

RuntimeFlags#isTestMode is needed now because it guards some non-IJ places in MPS
code base, I believe.
I cannot think of such usages for the headless flag.
All of them should go to some place like Evnironment or Platform which
are the entry points for MPS in the same way Application is the entry
point for IJ.