MPSI-17 Tooltip rendered for an ancestor node leads to a tool freeze
After some thinking it seems that I am unable to solve this issue
without changing the tooltips runtime.
Even if I change querylist runtime to respect the recursive
calls, the best I can offer is to replace all the tooltip cells that
have queries inside of them with empty cells (essentially throwing out
the query cells).
I checked that it is possible,yes.

Unfortunately this solution does not save us in the scenario, when
the tooltip cells contain query wrapped into some collection -- then
we will see the collections without queried cells in the pop-up, which might be odd.
Also another even worse scenario is when one wants to show the main cell of the
cell-with-tooltip in the pop-up using the query lang, is in a no way possible.
It is easy to see that the last scenario is not to be solved from the query lang.

The change appeared to be elementary, it is good to keep the
tooltips language as an extension which is generated into the editor language (almost true).