MPS-32697 Bundle statistics metadata with installer
It is also require to set customProductCode: in BuildContextImpl constructor applicationInfo.productCode will be set to productProperties.customProductCode. And this value will later be used in StatisticsRecorderBundledMetadataProvider#appendProductCode to get correct FUS json file.

Problem is that applicationInfo from Idea Community is used so applicationInfo.productCode == IC what is incorrect.

To avoid this MPS should have its own module with correct MPSApplicationInfo.xml