MPSSPRT-297 Custom RCP Build cannot have + character in its name
The problem in the issue is about plus signs.
We use File#getURL to encode URLs and Files#fromURL to decode URL.
Here the plus sign stays as is in URL and in Files#fromURL we used
URLDecoder which transforms '+' into space.
URLDecoder is intended for use in html urls which is not what we have
here.
Why do we need encoding/decoding when we uses URL only as a universal
storage format for paths? -- the alternative is to parse the URL string
manually and having URL just as a string wrapper which does not seem ok
at all.

The correct way is to use URI constructor with several arguments in
order to encode the string and use the single string constructor of URI
to decode the string.
See https://stackoverflow.com/questions/4571346/how-to-encode-url-to-avoid-special-characters-in-java/4571518#4571518, or the javadoc for URI.
It is correct and we did not have to use any utility methods for decoding-encoding.
Another source to work with jar urls:
https://stackoverflow.com/questions/402683/how-do-i-get-just-the-jar-url-from-a-jar-url-containing-a-and-a-specific-fi/402771#402771