build lang: relax dependency to generator error
Generally, it's indeed suspicious for a solution to depend on generator.
However, there are legitimate cases like migrations that need to establish a reference to a template and, therefore, need dependency to generator module.
I feel ModuleChecker is a wrong place for the check (especially if dependency has been 'derived' from module xml), but don't want to deal with it now.