MPS-32977 Handle IDEA's ControlFlowException
LOG.error(throwable) for a ProcessCanceledException from somewhere deep IDEA internals gives users feeling of broken MPS, while it's nothing but a IDEA's odd way to control application execution flow.
Unlike ActionDispatcher, I don't feel it's appropriate to propage control flow exceptions here up to caller, at least in 20.3 bugfix as it might lead to other hard-to-predict changes in functionality. Therefore, keep handling them here, albeit without logging for an end user.
I feel it's worth for MPS developers to be aware of the issue, that's why we go on logging it in an internal (development) model. After all, ability of MPS developers to withstand a lot of exceptions is widely renowned.