MPS-33007 windows problem with urls, paths, jars and slashes
VirtualFile from IJ returns normal path like C:/a/b/c/...
However our code in Files#fromURL instead retrieves the path
/C:/a/b/c... which is quite logical, but creates two non-identical IdeaFiles in MPS pointing
to the same location, leading to such problems as MPS-33007.

So our trick with path -> url -> path has given birth to another problem.
Thus the ugly fix of MPS-28009 is still needed because of that.
I removed that as a part of complex refactoring (MPSSPRT-297), and I was
wrong.
I have returned the fix back, just to make the problem disappear.

Though I see now that URL protocol as a mediator for our file paths in java API was a
really bad decision made by me, so it will be replaced in master.

Here is another discussion on this particular problem.
https://stackoverflow.com/questions/6164448/convert-url-to-normal-windows-filename-java
Though the post does not consider jar-file URLs.