MPS-32786 VCS Annotate bugfix.
Sometimes icons on the editor panel left highlighter were messed up
with annotate lines. The order of some UI events is changed in the fix.
The left highlighter now is relayouted and repainted directly after the
annotation update. Also, an additional bug with annotated cell
messages is fixed - after fixing the original bug there were duplicated
messages in the tooltips.