Drop tracking of typeof() invocations from inference rules, MPS-33050
The incremental typechecking algorithm used to divide dependencies into
two categories: those which affect the types, and those that don't.
This does very little in terms of optimization, but adds a layer of complexity
and causes bugs that are really difficult to find.

It's safe to assume that a type inference rule applied to a node does
affect the node's type, so this distinction is really unnecessary.