Replacing URL/URI with Path
Changing IFile classes in a minimalistic way in order to provide the
mapping IFile -- Path -- String instead of IFile -- URI -- String,
namely get rid of Files#fromURL.

We have had a lot of problems caused by the fact that URL/URI are
not _that_ good to support archive paths and unix/windows formats.

It seems that even having plain strings would be easier and cheaper for
us.
Instead here we start using a path abstraction
(jetbrains.mps.vfs.Path) which lies at a lower level than IFile, since
it does not address the physical file system.
Furthermore, it seems to be a right thing to do since the code is cluttered with
ad hoc path construction and different parts of our api pass the paths
as strings.
TODO plan is outlined in the javadoc of jetbrains.mps.vfs.Path.