trove.test: bring unused collections.trove lang back to make sure its RT is in CP
Explicit commit not to get into this trouble again, when removing an unused language renders unexpected compilation errors.
Well, not to prevent, but rather to discover faster than this time.
It's odd that collections.trove language have to be 'used' this way. Perhaps, coll.trove shall re-export trove.jar dependency (as a solution?)