MPS-32549 hot-fixed:
The fix is twofold:
1. fallback to the existing file if vcs does not respond properly.
This presumably fixes all the problems in IDE.
2. to return null instead of loaded model in case there are zero
substreams in the data source. This fixes the failing
'delete-and-restore' test.

The real reason is that the system is not synchronized with ij vcs,
presumably we need to invoke at least update in the worker thread of
ChangeListManager.
The delete-and-restore test fails because our update procedure is
happens independently of the ChangeListManager state.