MPS-32827 VCS History using commits graph
Current annotation algorithm uses the list of revisions for the file obtained
from the server. The changes for the root are calculated between two adjacent
revisions. It may happen that these revisions are from two different branches.
In this case the obtained changes make no sense. Merge revisions are not handled
correctly as well. The changes from the older revisions can be erroneously
displayed as changes from the newer merge revision. Therefore, the revisions
graph should be used for calculating the correct history of the changes.