Generator: improve GP declaration to address tests scenario
With entries to consume broad range of languages in a single step,
(1) introduce a mechanism to encapsulate each contribution into distinct transform step (to become an option),
(2) facilitate 'consumption' of an engaged generator so that subsequent step don't have a chance to include it as well.
Check new devkit/gp for tests against one of BL test solutions.