vcs diff: fixed a problem with the opposite change set building.
The opposite change set was not actually filled with the model changes
while building the main change set since the "commit" operation for
the main change set was called after building the opposite change set.

Now we first fill in the model changes for the main change set and only
then build the opposite change set.

Besides, the opposite change set was created anew for each rebuild of
the main change set. This is also fixed. The opposite change set is
now cleared and filled in rather than created anew during the rebuild
operation.