MPSSPRT-323 Pre-bundled MPS plugins should be loaded when executing tests outside of the same process
We will be moving towards the IJ configuration for the plugins to
enable.
First of all when running tests fromr ide/ant we want to have all the
installed plugins unless the system property idea.load.plugins.id is
already set.

The property idea.load.plugins.id might be empty which means no
restrictions on the plugins in IJ.
We allow this by introducing another option in EnvironmentConfig.

Secondly, we remove all the hacks connected with the listing all the
bundled plugins.
IJ removed error if the plugin is duplicated in cp and in the property,
so we do not need this logic.