MPS-32683 copy roots for registered module
copyModelsAndFacets used to work with a module registered in a repo. Only language attached to a repo has generators 'revalidated' from module descriptor (unless explicitly initialized; that wasn't the case of MRF.instantiateModule anyway).
I feel the approach is flawed, and we'd rather make a copy on descriptor level, rather than deal with attached module instances to deal with model roots. However, it's bigger change compared to module registered to the project repo (as it used to be prior to the change) AND exposed in a project.