MPS-33036 A fix for missing gutter marks.
Gutter marks were missing for change messages in merge dialog
and also if a diff dialog was opened from the push dialog. Both dialog
windows are modal dialogs. Probably, because of this a queue in
ChangeGroupMessages was blocked. Making this queue to work in the threads
other than EDT fixes the problem.