Ensure checking rules's quickfixes are applied correctly, MPS-33221
Introduce another flavour to track quickfix provider, which is the object
created by the rule that corresponds to the particular application (instance)
of a quickfix. Without this flavour a quickfix is only identified by the pair
node-rule.

Also enable execution of quickfix code in case the editor doesn't
have single selection at the moment of application.