Attempt #2 MPSSPRT-316
Explicit api to provide dependent nodes for a scope.
The next step for it is to be extracted into scopes dsl

Changing the logic of finding the invalidated nodes by using transitive
closure rather than just neighbours in the many2many map.
When we listen to read we have no choice, but when explicitly providing
the dependencies, we can save some effort by doing additional work at
runtime.