Introduce aux data to TypecheckingProvider. Refactor typechecking facade
Aux data container is introduced to typechecking provider
to enable accessing of provider-specific data independently
of the current typechecking session, which may not be available
to the client. This data is to be cleared after controller
is popped off the controller stack. It can be shared between
instances of "incremental" queries.