ref scope: Mechanism to share evaluation context
ReferenceScopeHelper is not the perfect place, but it's already responsible for reference scope, and it's reasonable to keep evaluation context there. Although might be better to have both scope and context caching attached to ModelAccess rather than to a repository (to stress it's bound to access time, rather than module container).