LanguageRuntime: extract code to collect runtime modules
To address bootstrap scenario in an identifiable manner, code to extract language runtime modules based either on source module or deployed LR instance.
It's confusing to have it as part of SLanguage impl (generally, implies uses of deployed info). Don't want to keep it burried deep down in RuntimesOfUsedLanguageCalculator as I intend to re-use the code once visitor is in place.