MPS-33610 Fixing memory leaks in VCS.
First of all, ChangeGroupMessages#dispose method was not called when objects of this class were used in
Diff and Merge dialogs. Now it's fixed. Second, instances of MergingUpdateQueue class should be disposed using
Disposer#dispose, rather than by direct call of MergingUpdateQueue#dispose. This is a common rule for all
implementations of Disposable interface. Same problem was fixed in TripleChangeGroupLayout and EditorAnnotation.

Last thing, I've got another problem on MPS shutdown after fixing the bugs above.
CurrentDifferenceRegistry#getCurrentDifference method was used to get an existing instance of CurrentDifference
for a given model. However, if CurrentDifferenceRegistry is already disposed, the instance of CurrentDifference is
not found and it is constructed anew. The fix is to store the instances of CurrentDifference and use them at dispose
step.