MPSSPRT-343, MPSSPRT-333 fix is provided
The main fix is to remove the attempt to support VirtualFileImpl in MPSNodeEditorProvider#accept.
It did not work and led to the freeze, which is described in MPSSPRT-343.

A possible motivation to stick to MPSNodeVirtualFiles in our editors is
as follows.
Since it is impossible to map one-to-one disk files (VirtualFileImpl) to MPSNodeVirtualFiles (roots), one consistent way to provide editors for different files for MPS is to rewrite all the callers of the platform #openEditor methods so that instead of VirtualFileImpl we pass MPSNodeVirtualFile.

For example, we can write our own NavigationBar, as we already did in IdeaPlugin (see MPSNavBarExtension).
Or with VCS tool window, we could try either creating an action, which would open a corresponding file for the changed file; or we could write another view altogether, listing not the changed files but the changed models, roots and nodes (so MPS level, a level higher than the plain disk level).

Other than those, we have to stick to the correspondence "MPSNodeVirtualFile -> MPS view, managed by MPS editor and VirtualFileImpl -> plain text view, managed by Intellij"