MPSSPRT-344 Error tooltips don't show up in MPS 2021.1
In revision 97b7fdbeb0c0032ebe385482023bd515a671fe01 the method EditorComponent#getEditorMessagesFor method
was changed in order to better display VCS Annotation messages for the annotated editor. The tooltip
calculation for the instances of HighlighterMessage messages was broken, so that it stopped to work for
multiline cell collections. Here, this calculation is repaired and VCS Annotation messages seem to work fine
 either.