ConsoleContext: refactor smodel.query dependency
The only use of CC as QueryExecutionContext is BLCommand.execute with nested QueryExpression. These can be handled with additional with statement.
I don't like implementation of with/query that assume 'context' variable name, might want to look into that later.
I don't expect any code to get affected by renamed of ConsoleContext parameter as I'm unaware of a mechanism to access it from BLCommand statements.
Don't see a reason to use ProjectScope from findUsages when Project.getScope gives scope I would assume to be the perfect fit here.
With this change, no more uses of [squery-runtime] in IDEA-managed code