Proper ant::migrate task
Use endorsed mechanism to pass task-->worker params. Script is generic class, shall not bear arguments specific to the task. Although startup properties is not the best way to go, it's easy to fix implementation when the logic is well-confined.
Facilitate migration of multiple projects (both with custom <project path> and <dirset>)