Evaluate when_concrete's body in correct context, MPS-33051
The body of a when_concrete statement is evaluated in the
context of a node that initiated rules application.
So, any read operations that are caught in order to keep track
of dependencies in the incremental mode are recorded
against that node, whereas they are expected to be recorded
against the node that initiated adding of the when_concrete
block.

This patch introduces ContextRunnable that ensures the body
of a when_concrete statement is evaluated in the correct
context.