Generator: private ML + runtime
Some labels are useless outside of a generator and either waste space in checkpoint models or make these models change on each generation if the labeled node shows up at a step prior to checkpoint.
The idea is to denote such LMs as 'private' and don't serialize them into checkpoint models.