MPSSPRT-295 smart tootips in the editor
The editor tooltips of the new type are implemented. They are based on the IDEA’s
classes from com.intellij.codeInsight.hint. The com.intellij.openapi.editor.Editor interface
was implemented to fulfil the platform API. However, this is a temporary solution. We should
get rid of this implementation and find a better way to reuse the platform code.

New tooltips have the following advantages:
- it is possible to click on links inside the tooltip
- the tooltip is always located inside the application main window
- it is possible to scroll the text inside the tooltip if the text is too long
- independent parts of the tooltip can be separated with a nice thin line