SModule: account for getModels() calls prior to attach()
There are hacky checks for changed models that happen in repo.registerModule and in doUpdateModelRoots().
With models not available until module is attached and its model roots are initialized, no reason to ask for models and even if this happens, have to account models have not been properly initialized (INCOMPLETE), denote them as such.