Refresh CP model for structure aspect
With recent change in generator, 40241c60, template reference to a node created from Create Root rule is restored as a static reference (we consult null input now in addition to input history when restoring ReferenceInfo_Template).
Prior to the change, LanguageConceptSwitch reference was first added as a dynamic one, and then replaced with a static. Replacement of a reference happens with (remove, add) sequence, that's why StaticFieldReference.classifier, which comes first in template, happened to be last in the CP model (added first, then followed by VariableReference.variableDeclaration, then the first one removed and another one, static, added).