MPS-29786 AE from DynamicReference
Re-use existing node<Type> instead of making a copy of a node<ClassifierType> from a detached node. ClassifierType.classifier may be a dynamic reference which doesn't like to get resolved from a detached node.
In fact, AdditionalForLoopVariable.type is of no use, there's custom typeof rule that always takes type of the first variable. Prior to MPS-14335 fix, the type was hard-coded <int>, which might not necessarily be a poor solution.