BL: facility to match node<Classifier> vs its stub
Mostly a refactoring for BaseLanguageEnvironmentHelper logic, which was bound to MPS-supplied stub solutions.
The reason for that limitation seems to be poor performance. Tried to mitigate with cached set of modules to consult for stub models.
Decided to keep utilities close to the language module, although this might not necessarily be right solution.
My motivation is to reduce number of modules in BL compilation cycle; with utilities kept in distinct module (like bl.util), it's easier to drag more dependencies than necessary. However, this particular fix doesn't help to get rid of bl.util module dependency, just reduce a cohesion.
Aah, yes, I didn't like ClassType use.